chore(tests): Bring tests in line with src project structure
Move from monolithic `Tsk.Tests` to separate:
- `tests/Tsk.CLI.Tests/`
- `tests/Tsk.Domain.Tests/`
- `tests/Tsk.Infrastructure/`

Also adjust namespacing to reflect new struture (e.g.
`Tsk.Domain.Tests.Entities`)
