+++ tests/Tsk.Infrastructure.Tests/Tsk.Infrastructure.Tests.csproj
@@ -8,7 +8,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Bogus" Version="35.6.3" />
     <PackageReference Include="coverlet.collector" Version="6.0.2" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.12.0" />
     <PackageReference Include="Moq" Version="4.20.72" />
@@ -21,9 +20,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Tsk.Domain\Tsk.Domain.csproj" />
     <ProjectReference Include="..\..\src\Tsk.Infrastructure\Tsk.Infrastructure.csproj" />
-    <ProjectReference Include="..\..\src\Tsk.CLI\Tsk.CLI.csproj" />
+    <ProjectReference Include="..\Tsk.TestSupport\Tsk.TestSupport.csproj" />
   </ItemGroup>
 
 </Project>
